Fjerne ting vi ikke skal bruke men som henger igjen i testdata, der er det greit at de ligger som extension functions
