+++ libs/saksbehandling-common/src/main/kotlin/person/Folkeregisteridentifikator.kt
@@ -5,7 +5,6 @@ import com.fasterxml.jackson.annotation.JsonValue
 import no.nav.etterlatte.libs.common.feilhaandtering.UgyldigForespoerselException
 import no.nav.etterlatte.libs.common.feilhaandtering.krevIkkeNull
 import no.nav.etterlatte.libs.common.logging.sikkerlogger
-import java.time.LocalDate
 
 /**
  * Norwegian national identity number
@@ -50,72 +49,12 @@ class Folkeregisteridentifikator private constructor(
     }
 
     /**
-     * @return birthdate as [LocalDate]
-     * Denne er feil ref https://nav-it.slack.com/archives/C020Q1K6EAY/p1727427982821959?thread_ts=1727351673.938029&cid=C020Q1K6EAY
-     */
-    @Deprecated("Denne blir feil for alle fødselsnummere")
-    fun getBirthDte(): LocalDate {
-        val fnrMonth = value.slice(2 until 4).toInt()
-
-        val fnrDay = value.slice(0 until 2).toInt()
-        val day = if (isDNumber()) fnrDay - 40 else fnrDay
-
-        val month =
-            if (isTestNorgeNumber()) {
-                fnrMonth - 80
-            } else if (isHNumber()) {
-                fnrMonth - 40
-            } else {
-                fnrMonth
-            }
-
-        return LocalDate.of(getYearOfBirth(), month, day)
-    }
-
-    /**
      * Checks if the identity number is of type D-number.
      *
      * A D-number consists of 11 digits, of which the first six digits show the date of birth,
      * but the first digit is increased by 4.
      */
     fun isDNumber(): Boolean = Character.getNumericValue(value[0]) in 4..7
-
-    /**
-     * Calculates year of birth using the individual number.
-     *
-     * @return 4 digit year of birth as [Int]
-     */
-    @Deprecated("Denne er feil og må fases ut til fordel for å hente foedselsdato direkte fra PDL")
-    private fun getYearOfBirth(): Int {
-        val year = value.slice(4 until 6).toInt()
-        val individnummer = value.slice(6 until 9).toInt()
-
-        return firesifretAarstallFraTosifret(year, individnummer)
-    }
-
-    /**
-     * Sjekker om fødselsnummeret er av typen "Hjelpenummer".
-     *
-     * H-nummer er et hjelpenummer, en virksomhetsintern, unik identifikasjon av en person som
-     * ikke har fødselsnummer eller D-nummer eller hvor dette er ukjent.
-     *
-     * Brukes også for identer i test som er opprettet som "NAV syntetisk" i Dolly
-     */
-    private fun isHNumber(): Boolean = Character.getNumericValue(value[2]) in 4..7
-
-    /**
-     * Sjekker om fødselsnummeret er av typen "Syntetisk bruker fra Skatteetaten".
-     */
-    private fun isTestNorgeNumber(): Boolean = Character.getNumericValue(value[2]) >= 8
-
-    /**
-     * Sjekker om fødselsnummeret er av typen "Felles Nasjonalt Hjelpenummer".
-     *
-     * Brukes av helsevesenet i tilfeller hvor de har behov for unikt å identifisere pasienter
-     * som ikke har et kjent fødselsnummer eller D-nummer.
-     */
-
-    private fun isFhNumber(): Boolean = Character.getNumericValue(value[0]) in 8..9
 
     override fun equals(other: Any?): Boolean = this.value == (other as? Folkeregisteridentifikator)?.value
